+++ app/src/main/java/app/revanced/manager/di/ViewModelModule.kt
@@ -23,4 +23,5 @@ val viewModelModule = module {
     viewModelOf(::InstalledAppInfoViewModel)
     viewModelOf(::UpdatesSettingsViewModel)
     viewModelOf(::BundleListViewModel)
+    viewModelOf(::ChangelogsViewModel)
 }