+++ app/src/main/java/app/revanced/manager/ui/model/navigation/Nav.kt
@@ -83,6 +83,9 @@ object Settings {
     data object About : Destination
 
     @Serializable
+    data object Changelogs : Destination
+
+    @Serializable
     data object Contributors : Destination
 
     @Serializable