+++ app/src/main/java/app/revanced/manager/ui/screen/settings/DeveloperSettingsScreen.kt
@@ -12,14 +12,11 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.res.stringResource
 import app.universal.revanced.manager.R
-import app.revanced.manager.domain.manager.PreferencesManager
 import app.revanced.manager.ui.component.AppTopBar
 import app.revanced.manager.ui.component.GroupHeader
-import app.revanced.manager.ui.component.settings.BooleanItem
 import app.revanced.manager.ui.component.settings.SettingsListItem
 import app.revanced.manager.ui.viewmodel.DeveloperOptionsViewModel
 import org.koin.androidx.compose.koinViewModel
-import org.koin.compose.koinInject
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -28,7 +25,6 @@ fun DeveloperSettingsScreen(
     vm: DeveloperOptionsViewModel = koinViewModel()
 ) {
     val scrollBehavior = TopAppBarDefaults.pinnedScrollBehavior(rememberTopAppBarState())
-    val prefs: PreferencesManager = koinInject()
 
     Scaffold(
         topBar = {
@@ -41,14 +37,7 @@ fun DeveloperSettingsScreen(
         modifier = Modifier.nestedScroll(scrollBehavior.nestedScrollConnection),
     ) { paddingValues ->
         Column(modifier = Modifier.padding(paddingValues)) {
-            GroupHeader(stringResource(R.string.manager))
-            BooleanItem(
-                preference = prefs.showDeveloperSettings,
-                headline = R.string.developer_options,
-                description = R.string.developer_options_description,
-            )
-
-            GroupHeader(stringResource(R.string.patches))
+            GroupHeader(stringResource(R.string.patch_bundles))
             SettingsListItem(
                 headlineContent = stringResource(R.string.patches_force_download),
                 modifier = Modifier.clickable(onClick = vm::redownloadBundles)